@@ -1,22 +1,20 @@
 package com.baeldung.jackson.annotation.deserialization;
 
-import java.io.IOException;
-
 import com.baeldung.jackson.annotation.dtos.ItemWithSerializer;
 import com.baeldung.jackson.annotation.dtos.User;
-import com.fasterxml.jackson.core.JsonParser;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.DeserializationContext;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.deser.std.StdDeserializer;
-import com.fasterxml.jackson.databind.node.IntNode;
+import tools.jackson.core.JsonParser;
+import tools.jackson.core.JacksonException;
+import tools.jackson.databind.DeserializationContext;
+import tools.jackson.databind.JsonNode;
+import tools.jackson.databind.deser.std.StdDeserializer;
+import tools.jackson.databind.node.IntNode;
 
 public class ItemDeserializerOnClass extends StdDeserializer<ItemWithSerializer> {
 
     private static final long serialVersionUID = 5579141241817332594L;
 
     public ItemDeserializerOnClass() {
-        this(null);
+        super(ItemWithSerializer.class);
     }
 
     public ItemDeserializerOnClass(final Class<?> vc) {
@@ -27,8 +25,8 @@ public class ItemDeserializerOnClass extends StdDeserializer<ItemWithSerializer>
      * {"id":1,"itemNr":"theItem","owner":2}
      */
     @Override
-    public ItemWithSerializer deserialize(final JsonParser jp, final DeserializationContext ctxt) throws IOException, JsonProcessingException {
-        final JsonNode node = jp.getCodec()
+    public ItemWithSerializer deserialize(final JsonParser jp, final DeserializationContext ctxt) throws JacksonException {
+        final JsonNode node = jp.objectReadContext()
             .readTree(jp);
         final int id = (Integer) ((IntNode) node.get("id")).numberValue();
         final String itemName = node.get("itemName")
