@@ -1,15 +1,15 @@
 package com.baeldung.jackson.ignore;
 
-import com.fasterxml.jackson.core.JsonGenerator;
-import com.fasterxml.jackson.core.JsonParseException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.SerializerProvider;
-import com.fasterxml.jackson.databind.ser.BeanPropertyWriter;
-import com.fasterxml.jackson.databind.ser.FilterProvider;
-import com.fasterxml.jackson.databind.ser.PropertyFilter;
-import com.fasterxml.jackson.databind.ser.PropertyWriter;
-import com.fasterxml.jackson.databind.ser.impl.SimpleBeanPropertyFilter;
-import com.fasterxml.jackson.databind.ser.impl.SimpleFilterProvider;
+import tools.jackson.core.JsonGenerator;
+import tools.jackson.databind.ObjectMapper;
+import tools.jackson.databind.SerializationContext;
+import tools.jackson.databind.ser.BeanPropertyWriter;
+import tools.jackson.databind.ser.FilterProvider;
+import tools.jackson.databind.ser.PropertyFilter;
+import tools.jackson.databind.ser.PropertyWriter;
+import tools.jackson.core.exc.StreamReadException;
+import tools.jackson.databind.ser.std.SimpleBeanPropertyFilter;
+import tools.jackson.databind.ser.std.SimpleFilterProvider;
 import org.junit.Test;
 
 import java.io.IOException;
@@ -21,7 +21,7 @@ import static org.junit.Assert.assertThat;
 public class IgnoreFieldsWithFilterUnitTest {
 
     @Test
-    public final void givenTypeHasFilterThatIgnoresFieldByName_whenDtoIsSerialized_thenCorrect() throws JsonParseException, IOException {
+    public final void givenTypeHasFilterThatIgnoresFieldByName_whenDtoIsSerialized_thenCorrect() throws StreamReadException, IOException {
         final ObjectMapper mapper = new ObjectMapper();
         final SimpleBeanPropertyFilter theFilter = SimpleBeanPropertyFilter.serializeAllExcept("intValue");
         final FilterProvider filters = new SimpleFilterProvider().addFilter("myFilter", theFilter);
@@ -39,23 +39,23 @@ public class IgnoreFieldsWithFilterUnitTest {
     }
 
     @Test
-    public final void givenTypeHasFilterThatIgnoresNegativeInt_whenDtoIsSerialized_thenCorrect() throws JsonParseException, IOException {
+    public final void givenTypeHasFilterThatIgnoresNegativeInt_whenDtoIsSerialized_thenCorrect() throws StreamReadException, IOException {
         final PropertyFilter theFilter = new SimpleBeanPropertyFilter() {
             @Override
-            public final void serializeAsField(final Object pojo, final JsonGenerator jgen, final SerializerProvider provider, final PropertyWriter writer) throws Exception {
+            public final void serializeAsProperty(final Object pojo, final JsonGenerator jgen, final SerializationContext provider, final PropertyWriter writer) throws Exception {
                 if (include(writer)) {
                     if (!writer.getName()
                         .equals("intValue")) {
-                        writer.serializeAsField(pojo, jgen, provider);
+                        writer.serializeAsProperty(pojo, jgen, provider);
                         return;
                     }
 
                     final int intValue = ((MyDtoWithFilter) pojo).getIntValue();
                     if (intValue >= 0) {
-                        writer.serializeAsField(pojo, jgen, provider);
+                        writer.serializeAsProperty(pojo, jgen, provider);
                     }
-                } else if (!jgen.canOmitFields()) { // since 2.3
-                    writer.serializeAsOmittedField(pojo, jgen, provider);
+                } else if (!jgen.canOmitProperties()) { // since 2.3
+                    writer.serializeAsOmittedProperty(pojo, jgen, provider);
                 }
             }
 
