@@ -9,17 +9,16 @@ import java.io.IOException;
 import org.junit.Test;
 
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
-import com.fasterxml.jackson.core.JsonGenerator;
-import com.fasterxml.jackson.core.JsonParseException;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.SerializerProvider;
-import com.fasterxml.jackson.databind.ser.BeanPropertyWriter;
-import com.fasterxml.jackson.databind.ser.FilterProvider;
-import com.fasterxml.jackson.databind.ser.PropertyFilter;
-import com.fasterxml.jackson.databind.ser.PropertyWriter;
-import com.fasterxml.jackson.databind.ser.impl.SimpleBeanPropertyFilter;
-import com.fasterxml.jackson.databind.ser.impl.SimpleFilterProvider;
+import tools.jackson.core.JacksonException;
+import tools.jackson.databind.ObjectMapper;
+import tools.jackson.databind.json.JsonMapper;
+import tools.jackson.databind.ser.BeanPropertyWriter;
+import tools.jackson.databind.ser.FilterProvider;
+import tools.jackson.databind.ser.PropertyFilter;
+import tools.jackson.databind.ser.PropertyWriter;
+import tools.jackson.core.exc.StreamReadException;
+import tools.jackson.databind.ser.std.SimpleBeanPropertyFilter;
+import tools.jackson.databind.ser.std.SimpleFilterProvider;
 
 public class JacksonSerializationIgnoreUnitTest {
 
@@ -28,7 +27,7 @@ public class JacksonSerializationIgnoreUnitTest {
     // ignore
 
     @Test
-    public final void givenOnlyNonDefaultValuesAreSerializedAndDtoHasOnlyDefaultValues_whenSerializing_thenCorrect() throws JsonParseException, IOException {
+    public final void givenOnlyNonDefaultValuesAreSerializedAndDtoHasOnlyDefaultValues_whenSerializing_thenCorrect() throws StreamReadException, IOException {
         final ObjectMapper mapper = new ObjectMapper();
         final String dtoAsString = mapper.writeValueAsString(new MyDtoIncludeNonDefault());
 
@@ -37,7 +36,7 @@ public class JacksonSerializationIgnoreUnitTest {
     }
 
     @Test
-    public final void givenOnlyNonDefaultValuesAreSerializedAndDtoHasNonDefaultValue_whenSerializing_thenCorrect() throws JsonParseException, IOException {
+    public final void givenOnlyNonDefaultValuesAreSerializedAndDtoHasNonDefaultValue_whenSerializing_thenCorrect() throws StreamReadException, IOException {
         final ObjectMapper mapper = new ObjectMapper();
         final MyDtoIncludeNonDefault dtoObject = new MyDtoIncludeNonDefault();
         dtoObject.setBooleanValue(true);
@@ -49,7 +48,7 @@ public class JacksonSerializationIgnoreUnitTest {
     }
 
     @Test
-    public final void givenFieldIsIgnoredByName_whenDtoIsSerialized_thenCorrect() throws JsonParseException, IOException {
+    public final void givenFieldIsIgnoredByName_whenDtoIsSerialized_thenCorrect() throws StreamReadException, IOException {
         final ObjectMapper mapper = new ObjectMapper();
         final MyDtoIgnoreFieldByName dtoObject = new MyDtoIgnoreFieldByName();
         dtoObject.setBooleanValue(true);
@@ -62,7 +61,7 @@ public class JacksonSerializationIgnoreUnitTest {
     }
 
     @Test
-    public final void givenFieldIsIgnoredDirectly_whenDtoIsSerialized_thenCorrect() throws JsonParseException, IOException {
+    public final void givenFieldIsIgnoredDirectly_whenDtoIsSerialized_thenCorrect() throws StreamReadException, IOException {
         final ObjectMapper mapper = new ObjectMapper();
         final MyDtoIgnoreField dtoObject = new MyDtoIgnoreField();
 
@@ -75,9 +74,10 @@ public class JacksonSerializationIgnoreUnitTest {
 
     // @Ignore("Jackson 2.7.1-1 seems to have changed the API for this case")
     @Test
-    public final void givenFieldTypeIsIgnored_whenDtoIsSerialized_thenCorrect() throws JsonParseException, IOException {
-        final ObjectMapper mapper = new ObjectMapper();
-        mapper.addMixIn(String[].class, MyMixInForIgnoreType.class);
+    public final void givenFieldTypeIsIgnored_whenDtoIsSerialized_thenCorrect() throws StreamReadException, IOException {
+        final ObjectMapper mapper = JsonMapper.builder()
+            .addMixIn(String[].class, MyMixInForIgnoreType.class)
+            .build();
         final MyDtoWithSpecialField dtoObject = new MyDtoWithSpecialField();
         dtoObject.setBooleanValue(true);
 
@@ -90,7 +90,7 @@ public class JacksonSerializationIgnoreUnitTest {
     }
 
     @Test
-    public final void givenNullsIgnoredOnClass_whenWritingObjectWithNullField_thenIgnored() throws JsonProcessingException {
+    public final void givenNullsIgnoredOnClass_whenWritingObjectWithNullField_thenIgnored() throws JacksonException {
         final ObjectMapper mapper = new ObjectMapper();
         final MyDtoIgnoreNull dtoObject = new MyDtoIgnoreNull();
 
@@ -103,9 +103,10 @@ public class JacksonSerializationIgnoreUnitTest {
     }
 
     @Test
-    public final void givenNullsIgnoredGlobally_whenWritingObjectWithNullField_thenIgnored() throws JsonProcessingException {
-        final ObjectMapper mapper = new ObjectMapper();
-        mapper.setSerializationInclusion(Include.NON_NULL);
+    public final void givenNullsIgnoredGlobally_whenWritingObjectWithNullField_thenIgnored() throws JacksonException {
+        final ObjectMapper mapper = JsonMapper.builder()
+            .changeDefaultPropertyInclusion(incl -> incl.withValueInclusion(Include.NON_NULL))
+            .build();
         final MyDto dtoObject = new MyDto();
 
         final String dtoAsString = mapper.writeValueAsString(dtoObject);
