@@ -1,10 +1,11 @@
 package com.baeldung.jackson.unknownproperties;
 
-import com.fasterxml.jackson.core.JsonParseException;
-import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.JsonMappingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.exc.UnrecognizedPropertyException;
+import tools.jackson.core.exc.StreamReadException;
+import tools.jackson.databind.DeserializationFeature;
+import tools.jackson.databind.DatabindException;
+import tools.jackson.databind.ObjectMapper;
+import tools.jackson.databind.json.JsonMapper;
+import tools.jackson.databind.exc.UnrecognizedPropertyException;
 import org.junit.Test;
 
 import java.io.IOException;
@@ -16,7 +17,7 @@ import static org.junit.Assert.assertThat;
 public class UnknownPropertiesUnitTest {
 
     @Test
-    public final void givenNotAllFieldsHaveValuesInJson_whenDeserializingAJsonToAClass_thenCorrect() throws JsonParseException, JsonMappingException, IOException {
+    public final void givenNotAllFieldsHaveValuesInJson_whenDeserializingAJsonToAClass_thenCorrect() throws StreamReadException, DatabindException, IOException {
         final String jsonAsString = "{\"stringValue\":\"a\",\"booleanValue\":true}";
         final ObjectMapper mapper = new ObjectMapper();
 
@@ -30,9 +31,11 @@ public class UnknownPropertiesUnitTest {
     // tests - json with unknown fields
 
     @Test(expected = UnrecognizedPropertyException.class)
-    public final void givenJsonHasUnknownValues_whenDeserializingAJsonToAClass_thenExceptionIsThrown() throws JsonParseException, JsonMappingException, IOException {
+    public final void givenJsonHasUnknownValues_whenDeserializingAJsonToAClass_thenExceptionIsThrown() throws StreamReadException, DatabindException, IOException {
         final String jsonAsString = "{\"stringValue\":\"a\",\"intValue\":1,\"booleanValue\":true,\"stringValue2\":\"something\"}";
-        final ObjectMapper mapper = new ObjectMapper();
+        final ObjectMapper mapper = JsonMapper.builder()
+            .configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, true)
+            .build();
 
         final MyDto readValue = mapper.readValue(jsonAsString, MyDto.class);
 
@@ -43,14 +46,15 @@ public class UnknownPropertiesUnitTest {
     }
 
     @Test
-    public final void givenJsonHasUnknownValuesButJacksonIsIgnoringUnknownFields_whenDeserializing_thenCorrect() throws JsonParseException, JsonMappingException, IOException {
+    public final void givenJsonHasUnknownValuesButJacksonIsIgnoringUnknownFields_whenDeserializing_thenCorrect() throws StreamReadException, DatabindException, IOException {
         final String jsonAsString = // @formatter:off
                 "{\"stringValue\":\"a\"," +
                 "\"intValue\":1," +
                 "\"booleanValue\":true," +
                 "\"stringValue2\":\"something\"}"; // @formatter:on
-        final ObjectMapper mapper = new ObjectMapper();
-        mapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
+        final ObjectMapper mapper = JsonMapper.builder()
+            .configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
+            .build();
 
         final MyDto readValue = mapper.readValue(jsonAsString, MyDto.class);
 
@@ -61,7 +65,7 @@ public class UnknownPropertiesUnitTest {
     }
 
     @Test
-    public final void givenJsonHasUnknownValuesButUnknownFieldsAreIgnoredOnClass_whenDeserializing_thenCorrect() throws JsonParseException, JsonMappingException, IOException {
+    public final void givenJsonHasUnknownValuesButUnknownFieldsAreIgnoredOnClass_whenDeserializing_thenCorrect() throws StreamReadException, DatabindException, IOException {
         final String jsonAsString = // @formatter:off
                 "{\"stringValue\":\"a\"," +
                 "\"intValue\":1," +
